requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.112.4
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
googleapis-common-protos==1.69.1
gradio==4.43.0
gradio_client==1.3.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.33.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.20
langchain-chroma==0.2.2
langchain-community==0.3.19
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.2.19
langgraph-checkpoint==1.0.9
langsmith==0.3.13
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.65.4
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
posthog==3.19.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.0
pydantic-settings==2.8.1
pydantic_core==2.23.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyprojroot==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.10
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
waitress==3.0.2
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
